Problem: There was a missing import of NSString.
The imports assumed certain placement of the tinycbor dependency, which wasn't
suitable for how it was setup for Ditto's usage.

The `NSJSONReadingFragmentsAllowed` option wasn't used when decoding JSON and so
individual values couldn't be decoded.

Solution: Add the import for NSString in cbortojson_nsstring.m.

Change imports to just `cbor.h`, where necessary.

Add the `NSJSONReadingFragmentsAllowed` JSON reading option to the instance of
JSON decoding.